+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `AdCounterLabel` now shows the ad position across multiple ad breaks scheduled at the same time (e.g. `Ad 2 of 3` instead of `Ad 1 of 1` for each)
 
+### Fixed
+
+- `VolumeController` can store a volume to restore of `0` in some cases, causing unmute to not restore an audible volume level
+
 ## [4.9.1] - 2026-02-24
 
 ### Fixed